+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixed rendering of status sections, some properties had no effect #180
+- Sync project propertes after phase change #196
 
 ### Added
 - Installation writes to output which user it is connected with #187